Add specular light

+++ OpenGLCourseApp/OpenGLCourseApp/Light.h
@@ -11,8 +11,8 @@ public:
 	Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, 
 		GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat diffIntensity);
 
-	void UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation, 
-		GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+	void UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+		GLuint diffuseIntensityLocation, GLuint directionLocation);
 
 	~Light();
